refactor: inline odbc_util module into connection/mod.rs
Move the shared ODBC helper functions from a separate odbc_util.rs
file into an inline module within connection/mod.rs. Update all
references from crate::connection::odbc_util:: to crate::connection::
