+++ remote-table/src/connection/dm/mod.rs
@@ -1,7 +1,7 @@
 use crate::connection::ODBC_ENV;
 use crate::connection::dm::buffer::{buffer_to_batch, build_buffer_desc};
 use crate::connection::dm::row::row_to_batch;
-pub(crate) use crate::connection::odbc_util::{
+pub(crate) use crate::connection::{
     ms_since_epoch, ns_since_epoch, seconds_since_epoch, us_since_epoch,
 };
 use crate::{
@@ -322,4 +322,4 @@ fn dm_type_to_remote_type(data_type: odbc_api::DataType) -> DFResult<DmType> {
 }
 
 // Epoch helpers (seconds_since_epoch, ms_since_epoch, us_since_epoch,
-// ns_since_epoch) are re-exported from `crate::connection::odbc_util` above.
+// ns_since_epoch) are re-exported from `crate::connection` above.
